refactor: move modules into new subdirectories, clean up CodeInjector
