+++ src/transforms/Magic.ts
